refactor: rename file to kebab-case standard
